implement feed upload base layout
