+++ app/src/main/res/drawable/background_feed_upload_button.xml
@@ -1,0 +1,5 @@
+<?xml version="1.0" encoding="utf-8"?>
+<selector xmlns:android="http://schemas.android.com/apk/res/android">
+	<item android:drawable="@drawable/background_feed_upload_button_enabled" android:state_enabled="true" />
+	<item android:drawable="@drawable/background_feed_upload_button_disabled" android:state_enabled="false" />
+</selector>
