+++ app/src/main/res/drawable/background_feed_upload_button_disabled.xml
@@ -1,0 +1,6 @@
+<?xml version="1.0" encoding="utf-8"?>
+<shape xmlns:android="http://schemas.android.com/apk/res/android"
+	android:shape="rectangle">
+	<corners android:radius="14dp" />
+	<solid android:color="@color/_E3E3E3" />
+</shape>
