+++ app/src/main/res/drawable/icon_feed_upload.xml
@@ -1,0 +1,7 @@
+<?xml version="1.0" encoding="utf-8"?>
+<layer-list xmlns:android="http://schemas.android.com/apk/res/android">
+	<item
+		android:width="18dp"
+		android:height="18dp"
+		android:drawable="@drawable/ic_dropdown" />
+</layer-list>
