--- app/src/main/res/layout/activity_misson_upload.xml
@@ -1,28 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<layout xmlns:android="http://schemas.android.com/apk/res/android"
-	xmlns:app="http://schemas.android.com/apk/res-auto">
-	
-	<data>
-		
-		<variable
-			name="viewModel"
-			type="org.positive.daymotion.presentation.upload.MissionUploadViewModel" />
-	</data>
-	
-	<androidx.constraintlayout.widget.ConstraintLayout
-		android:layout_width="match_parent"
-		android:layout_height="match_parent">
-		
-		<TextView
-			android:id="@+id/title"
-			android:layout_width="wrap_content"
-			android:layout_height="wrap_content"
-			android:text="MISSION UPLOAD"
-			android:textSize="40dp"
-			app:layout_constraintBottom_toBottomOf="parent"
-			app:layout_constraintEnd_toEndOf="parent"
-			app:layout_constraintStart_toStartOf="parent"
-			app:layout_constraintTop_toTopOf="parent" />
-	
-	</androidx.constraintlayout.widget.ConstraintLayout>
-</layout>
