+++ app/src/main/res/values/colors.xml
@@ -7,6 +7,7 @@
 	<color name="_F3F5FE">#F3F5FE</color>
 	<color name="_EEEEEE">#EEEEEE</color>
 	<color name="_EDEDED">#EDEDED</color>
+	<color name="_E3E3E3">#E3E3E3</color>
 	<color name="_CDCDCD">#CDCDCD</color>
 	<color name="_F8F9FB">#F8F9FB</color>
 	<color name="_B9B9B9">#B9B9B9</color>
